cleanup: moved method into ProxyUtils
