+++ src/main/java/com/github/webdriverextensions/ProxyUtils.java
@@ -9,6 +9,7 @@ import org.apache.http.client.CredentialsProvider;
 import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.settings.Proxy;
+import org.apache.maven.settings.Settings;
 
 public class ProxyUtils {
     public static HttpHost createProxyFromSettings(Proxy proxySettings) throws MojoExecutionException {
@@ -38,5 +39,29 @@ public class ProxyUtils {
             }
         };
         Authenticator.setDefault(authenticator);
+    }
+
+    public static Proxy getProxyFromSettings(Settings settings, String proxyId) throws MojoExecutionException {
+        if (settings == null) {
+            return null;
+        }
+
+        if (proxyId != null) {
+            for (Proxy proxy : settings.getProxies()) {
+                if (proxyId.equals(proxy.getId())) {
+                    return proxy;
+                }
+            }
+            throw new MojoExecutionException("Configured proxy with id=" + proxyId + " not found in settings.xml");
+        }
+
+        // Get active http/https proxy
+        for (Proxy proxy : settings.getProxies()) {
+            if (proxy.isActive() && ("http".equalsIgnoreCase(proxy.getProtocol()) || "https".equalsIgnoreCase(proxy.getProtocol()))) {
+                return proxy;
+            }
+        }
+
+        return null;
     }
 }
